Refactored and reintroduced the sc_create's python tests
